// src/modules/booking/domain/enums/booking-status.enum.ts
export enum BookingStatus {
  OPEN = 'open',
  PAID = 'paid',
  WAITING_FOR_PAYMENT = 'waiting_for_payment',
  PAYMENT_EXPIRED = 'payment_expired',
  CONFIRMED = 'confirmed',
  WAITING_FOR_REFUND = 'waiting_for_refund',
  REFUNDED = 'refunded',
  DECLINED = 'declined',
  WORK_IN_PROGRESS = 'work_in_progress',
  WAITING_FOR_CONFIRMATION = 'waiting_for_confirmation',
  COMPLETED = 'completed',
  BID = 'bid',
  CANCEL_BY_SYSTEM = 'cancel_by_system',
  REJECTED = 'rejected',
  BOOKED = 'booked',
  FINISHED = 'finished',
  CANCELED_BY_CUSTOMER = 'canceled_by_customer',
  CANCELED_BY_ARTISAN = 'canceled_by_artisan',
  OPEN_ISSUE = 'open_issue',
}